+++ lib/components/movie_screen/poster.dart
@@ -1,0 +1,21 @@
+import 'package:flutter/material.dart';
+
+class Poster extends StatelessWidget {
+  const Poster({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: const BoxDecoration(
+        image: DecorationImage(
+          image: AssetImage("assets/images/mario_filme.jpg"),
+          fit: BoxFit.cover,
+        ),
+        borderRadius: BorderRadius.vertical(
+          bottom: Radius.circular(20),
+        ),
+      ),
+      height: MediaQuery.of(context).size.height / 2.6,
+    );
+  }
+}
